requirements.txt
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.12
django-allauth==0.38.0
django-appconf==1.0.2
django-cors-headers==2.5.2
django-filter==2.1.0
django-haystack==2.8.1
django-imagekit==4.0.1
django-markdown-deux==1.0.5
djangorestframework==3.9.2
django-ckeditor==5.7.1
django-js-asset==1.2.2
emoji==0.4.5
httplib2==0.12.1
idna==2.8
jieba==0.39
Markdown==2.6.8
markdown2==2.3.7
mysqlclient==1.4.2.post1
oauth==1.0.1
oauth2client==4.1.3
oauthlib==3.0.1
pilkit==2.0
Pillow==5.4.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
Pygments==2.3.1
PyMySQL==0.9.3
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
urllib3==1.24.2
Whoosh==2.7.4

